pyproject.toml
[tool.poetry]
name = "sky4 oop"
version = "0.1.0"
description = ""
authors = ["Elizaveta Nasytko <Liz99246@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
